requirements.txt
streamlit
pandas
gtts
python-dotenv
huggingface_hub
sacremoses
transformers 
sentencepiece
torch
